Fix cyclic SVD to allow its use with shift-and-invert
Now the (explicit) cyclic matrix for SVD has explicit zeros along
the diagonal. Otherwise, operations such as MatShift fail with an error
of 'New nonzero at (0,0) caused a malloc!' preventing its use in
combination with shift-and-invert (command-line example at the end of
section 4.4 of the users guide).

This is a long standing problem, dating back to 3.3.