Enable configure.py to install Sowing
In prefix installs of PETSc, the bfort executable is not available. So if
SLEPc's configure has to build Fortran stubs then it must download and
build Sowing.

Currently, the sowing.git repo is cloned from bitbucket.org

Optionally the user can specify the --download-sowing option to force the
installation of Sowing even if it is already available from PETSc.