CISS: avoid problem when running in parallel with usest=0
By default, ST will use a direct solver (sequential), so setup will fail
in parallel unless an external parallel LU solver is specified. However,
if usest=0 this problem should be ignored since we are not going to use
ST anyway.

A dirty fix is to set the ST type to SHIFT. We should find a better fix.