BV: fix one case where the computation of ||x||_B did not check for the sign
In Gram-Schmidt orthogonalization, the normalization of the vector
was sometimes done with BVNormVec, but sometimes not. In this latter
case it was missing the code that prevents computing the square root
of a negative value (coming from z'*B*z with indefinite B).

The fix introduces a new private function BV_SafeSqrt() to be used in
all required places.

Reported-by: Mohammad Asghar