BV: add some support for rank-deficient input basis in BVOrthogonalize()
Gram-Schmidt should usually work because after orthogonalization a
numerically zero vector turns into a 'random' vector when it is
normalized. Still a check for exactly zero norm is necessary for special
cases.

Cholesky orthogonalization fails in xPOTRF. A workaround is to shift the
matrix by a small amount, although better algorithms exist for this. The
output satisfies the QR residual but columns of Q are not orthonormal.