SVD: detect zero vectors in one-sided Lanczos variants
One-sided variants may become unstable in ill-conditioned problems where
a vector computed by the recurrence is zero (e.g. the Lauchli problem
ex15.c). We must error out in this event, otherwise the returned
solution is wrong.

This problem dates back to slepc-3.5 (in the change from IP to BV). In
slepc-3.4 the Lauchli problem works with one-sided variants. The current
implementation also works if MGS is used.