DS: remove DSNormalize()
The rationale for this change is:
+ It is never used across SLEPc.
+ For eigenvectors it does not make sense, because DSVectors() should
return eigenvectors already normalized (otherwise it is a bug).
+ For the Q matrix it does not make sense because it is an orthogonal
matrix, with columns already normalized (no sense to normalize as X
taking into account complex conjugate eigenvalues).
+ The only instance where DSNormalize() would make sense is in GHEP
where the user wants eigenvectors normalized wrt the 2-norm and not the
B-norm, but this can be easily implemented in user code.