PEP: change the way linearization-based solvers handle initial vectors
All PEP solvers based on linearization are of Krylov type, and hence
only one initial vector will be relevant for the linearization. If PEP
receives one or more initial vectors, these are stacked together in a
single vector (filling the remaining blocks randomly up to the degree of
the polynomial), which is passed as initial vector to the linearlization.

Formerly the initial vectors were discarded unless enough vectors (i.e.
the degree) were provided.