Made DM work with EPS power
The SNES context may be changed when the current DM is passed
to the new nonlinear solver which is used to compute an initial
guess for the current solver.

The current solver may not work later on if the SNES context is
changed.

We use a DM shell to temporarily store the SNES context that
is restored to the current solver once the initial solver is
done.