Use an empty vector for the template vector in all solver classes
Solver classes such as EPS or SVD allocate the internal basis with
BVSetSizesFromVec() by passing a template vector. This template vector
is only used to determine the vector type and global and local sizes,
but it need not contain an internal array. Using an empty vector may
represent a non-negligible memory savings.