FN: fix two bugs in the evaluation of functions of complex Hermitian matrices
The first bug is a missing conjugation when forming V*f(Lambda)*V'

The second bug occurs when A is complex Hermitian and alpha is a complex
scalar with nonzero imaginary part. In that case we should not diagonalize
alpha*A, since it is not Hermitian. Instead we must diagonalize A and
evaluate f(alpha*Lambda).

Reported-by: Storm Weiner