BV: safeguard for when a Krylov space of full dimension is requested
This fixes Issue #18

The issue appears when building a Krylov subspace of order n+1, where n
is the problem dimension. In sequential runs this does not cause grief,
because the last generated vector is numerically zero and the solver goes
on. However, in parallel this border case might imply that, due to roundoff
error, not all processes invoke a collective operation necessary to
normalize the last vector:

      /* estimate |v'| from |v| */
      ierr = BV_SquareSum(bv,j,c,&sum);CHKERRQ(ierr);
      *norm = beta*beta-sum;
      if (*norm <= 0.0) {
        ierr = BV_NormVecOrColumn(bv,j,v,norm);CHKERRQ(ierr);
      } else *norm = PetscSqrtReal(*norm);

Reported-by: Igor Gayday