BV: recover previous behavior of BVMatMult()
Before this change, the BVMatMult() operation was always being done
column by column, because MATOP_MAT_MULT had been removed from PETSc. With
an addition to PETSc, now it is possible to always call a MATPRODUCT_AB
operation on a dense B, and it falls back to a trivial implementation if
the operation is not available.

Thanks-To: Stefano Zampini