BV: fix TSQRCHOL in parallel due to the user-defined reduction
This fix supersedes 111aa1791fa3aeb120cecc6f918ed2581c4c5e60

The issue was not exclusive of Intel MPI, but of any MPICH derivative.
The MPI_Allreduce() operation does not work for 2 processes. The reason
is that MPICH is using a different algorithm for that case, where both
processes receive the input matrices and compute the result
independently. The QR factorization is unique up to a change of signs,
so it produced different result in each process. Hence, the user-
defined reduction operation must be defined non-commutative