config: slecp4py: honor DESTDIR make var
Top-level makefile honors destination directory (which is a path where
files are to be staged, orthogonal to the prefix path). The destination
path is specified via DESTDIR variable: `make DESTDIR=foo install`.

But, before this commit, the makefile target `petsc4py-install` did not
honor DESTDIR and ends up staging files into /, which created
counter-intuitive behavior that's hard to work around in packaging
scripts.

If DESTDIR is not set, then installation command is unchanged, exactly
as it was before this commit.

Signed-off-by: Alexei Colin <acolin@isi.edu>