slepc4py: build against the SLEPc being built
This fixes an issue with packaging SLEPc+slepc4py. The issue is that the
slepc4py Python extension builds against SLEPc installed in the system.
This effectively requires SLEPc to be installed into the system before
the slepc4py can be built, which is a problem for packaging recipes that
want to `--with-slepc4py=1` and have to stage the files into a DESTDIR
directory (i.e. cannot write then into the final installation prefix).
In this case, SLEPc won't be into its final install location at the time
the top-level build reaches the slepc4py build.

This patch allows slepc4py to build against the SLEPc that is being
built, located in the staging directory (DESTDIR). This involves two
actions:

1. set SLEPC_DIR to DESTDIR/PREFIX in the makefile

2. for output paths that are to be used at runtime, strip DESTDIR
from SLEPC_DIR

Note that SlepcConfig needs a separate SLEPC_DESTDIR member because
the DESTDIR member belongs to the PetscConfig parent class and cannot
be set to SLEPc's destination directory.

PETSc suffers from the same issue (the patch is similar):
https://gitlab.com/petsc/petsc/-/issues/862
https://gitlab.com/petsc/petsc/-/merge_requests/3748

Signed-off-by: Alexei Colin <acolin@isi.edu>