config: slepc4py: option to suppress numpy check for cross-compilation
On some HPC clusters it is necessary to cross-compile because the build hosts
(login nodes) have a different architecture (or CPU family) from worker ndoes,
for example on ALCF Theta HPC cluster workers are Xeon Phi, so turning on
cpu-family optimizations turns the build into a cross-compilation. The current
./configure fails with SIGILL (Illegal instruction) when it tries to import
petsc4py, which had been cross-compiled and cannot run on the build host.

This patch simply adds to `./configure` a flag to suppress autodetection/check
operations that involve loading dependencies on the build host.

Tested that SLEPc builds fine with `--have-petsc4py=1` and `--with-arpack=1
--with-scalapack=1 --with-slepc4py=1` on a system where PETSc, petsc4py, numpy,
arpack, scalapack had all been cross-compiled, i.e. where these fail with
SIGILL:

	python -c 'from petsc4py import PETSc`
	python -c 'import numpy`

Signed-off-by: Alexei Colin <acolin@isi.edu>