Modify the early call to DSSetType() in all high-level solvers
This is a modification of !465

Instead of calling setdstype() at XXXGetDS(), call it at both
XXXSetFromOptions() and XXXSetUp(). Also, add a public function
XXXSetDSType() that could be called by the user if required.

This change avoids issues when the XXXGetDS() is called in user code,
directly or indirectly, for instance when changing the solver prefix.