Use a better norm in ex34 FormNorm
The L2 norm of the current SNES solution just does not seem to produce
stable solves. And what's more, when using PJFNK, we may be computing
a Bx norm that doesn't even correspond to the current SNES solution,
e.g. it is a perturbed vector. Purpose of this test is to make sure
norm callbacks are working, not developing smart alternative Bx
norms