BV: improve performance of BVDotVec() in case of using CUDA and MPI
Must be careful when doing the reduction of the computed coefficients.
The previous version used VecGetArray() on bv->buffer, which implied a
copy to CPU of the full BVBufferVec when only the first column is
relevant at this point.

In this new version we also consider a shortcut in the case of
GPU-aware MPI implementations.